requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
aiofiles==24.1.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
asyncpg==0.30.0
supabase==2.9.1

# AI / LangChain
langchain==0.3.7
langchain-google-genai>=2.0.4
google-generativeai==0.8.3  # manter versão instalada — funciona com Python 3.13
langchain-community==0.3.7

# Media Processing
PyMuPDF==1.24.13
Pillow==11.0.0
python-magic-bin==0.4.14; sys_platform == "win32"
python-magic==0.4.27;      sys_platform != "win32"

# Task Scheduling
apscheduler==3.10.4

# Redis
redis==5.2.0

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
orjson==3.10.11
websockets==13.1
pytz>=2023.3
croniter>=1.4.1
